+++ gaming_site/src/App.tsx
@@ -1,12 +1,10 @@
 import { Grid, GridItem, Show } from "@chakra-ui/react";
-import { GameGrid, GenreList, Navbar, PlatformSelector } from "./components";
-import { Genre } from "./hooks/useGenres";
 import { useState } from "react";
-import { Platform } from "./hooks/useGames";
+import { GameGrid, GenreList, Navbar, PlatformSelector } from "./components";
+import { GameQuery } from "./types";
 
 const App = () => {
-  const [selectedGenre, setSelectedGenre] = useState<Genre | null>(null);
-  const [selectedPlaform, setSelectedPlaform] = useState<Platform | null>(null);
+  const [gameQuery, setGameQuery] = useState<GameQuery>({} as GameQuery)
   return (
     <Grid
       templateAreas={{
@@ -24,19 +22,18 @@ const App = () => {
       <Show above="lg">
         <GridItem area="aside" paddingX={5}>
           <GenreList
-            selectedGenre={selectedGenre}
-            onSelectGenre={(genre) => setSelectedGenre(genre)}
+            selectedGenre={gameQuery.genre}
+            onSelectGenre={(genre) => setGameQuery({...gameQuery,genre})}
           />
         </GridItem>
       </Show>
       <GridItem area="main">
         <PlatformSelector
-          selectedPlatform={selectedPlaform}
-          onSelectPlatform={(platform) => setSelectedPlaform(platform)}
+          selectedPlatform={gameQuery.platform}
+          onSelectPlatform={(platform) => setGameQuery({...gameQuery,platform})}
         />
         <GameGrid
-          selectedPlatform={selectedPlaform}
-          selectedGenre={selectedGenre}
+          gameQuery={gameQuery}
         />
       </GridItem>
     </Grid>
