+++ gaming_site/src/hooks/useGenres.ts
@@ -1,10 +1,7 @@
+import { Genre } from "../types";
 import useData from "./useData";
 
-export interface Genre {
-  id: number;
-  name: string;
-  image_background: string;
-}
+
 
 const useGenres = () => useData<Genre>("/genres");
