+++ gaming_site/src/types/index.ts
@@ -1,0 +1,24 @@
+export interface Platform {
+  id: number;
+  name: string;
+  slug: string;
+}
+export interface Game {
+  id: number;
+  name: string;
+  rating: number;
+  background_image: string;
+  parent_platforms: { platform: Platform }[];
+  metacritic: number;
+}
+
+export interface Genre {
+  id: number;
+  name: string;
+  image_background: string;
+}
+
+export interface GameQuery {
+  genre: Genre | null;
+  platform: Platform | null;
+}
